feat: extract evaluation_algorithm projection test code into separate file
