+++ src/unit_tests/core_server/internal/evaluation/evaluation_algorithm/common.cpp
@@ -1,0 +1,87 @@
+#include <catch2/catch_test_macros.hpp>
+#include <catch2/matchers/catch_matchers_vector.hpp>
+#include <memory>
+
+#include "core_server/internal/ceql/cel_formula/formula/visitors/formula_to_logical_cea.hpp"
+#include "core_server/internal/ceql/query_transformer/annotate_predicates_with_new_physical_predicates.hpp"
+#include "core_server/internal/evaluation/enumeration/tecs/tecs.hpp"
+#include "core_server/internal/evaluation/evaluator.hpp"
+#include "core_server/internal/evaluation/predicate_evaluator.hpp"
+#include "core_server/internal/parsing/ceql_query/parser.hpp"
+
+namespace CORE::Internal::Evaluation::UnitTests {
+
+RingTupleQueue::Tuple add_event(RingTupleQueue::Queue& ring_tuple_queue,
+                                uint64_t event_type_id,
+                                std::string val1,
+                                int64_t val2) {
+  uint64_t* data = ring_tuple_queue.start_tuple(event_type_id);
+  char* chars = ring_tuple_queue.writer<std::string>(val1.size());
+  memcpy(chars, &val1[0], val1.size());
+  int64_t* integer_ptr = ring_tuple_queue.writer<int64_t>();
+  *integer_ptr = val2;
+  return ring_tuple_queue.get_tuple(data);
+}
+
+std::vector<std::pair<std::pair<uint64_t, uint64_t>,
+                      std::vector<RingTupleQueue::Tuple>>>
+enumerator_to_vector(tECS::Enumerator& enumerator) {
+  enumerator.reset();
+  std::vector<std::pair<std::pair<uint64_t, uint64_t>,
+                        std::vector<RingTupleQueue::Tuple>>>
+    out;
+  for (auto info : enumerator) {
+    out.push_back(info);
+  }
+  return out;
+}
+
+std::string output_to_string(tECS::Enumerator& enumerator) {
+  enumerator.reset();
+  std::string out = "";
+  for (std::pair<std::pair<uint64_t, uint64_t>,
+                 std::vector<RingTupleQueue::Tuple>> info : enumerator) {
+    out += "[" + std::to_string(info.first.first) + ","
+           + std::to_string(info.first.second) + "]" + ", {";
+    for (auto tuple : info.second) {
+      RingTupleQueue::Value<std::string_view> val1(tuple[0]);
+      RingTupleQueue::Value<std::int64_t> val2(tuple[1]);
+      if (tuple.id() == 0) {
+        out += "SELL " + std::string(val1.get()) + " "
+               + std::to_string(val2.get()) + " ";
+      }
+      if (tuple.id() == 1) {
+        out += "BUY " + std::string(val1.get()) + " "
+               + std::to_string(val2.get()) + " ";
+      }
+    }
+    out += "}\n";
+  }
+  return out;
+}
+
+bool is_the_same_as(RingTupleQueue::Tuple tuple,
+                    uint64_t event_type_id,
+                    std::string name,
+                    int64_t value) {
+  if (tuple.id() != event_type_id) {
+    return false;
+  }
+  std::string_view
+    tuple_name = RingTupleQueue::Value<std::string_view>(tuple[0]).get();
+  int64_t tuple_val = RingTupleQueue::Value<int64_t>(tuple[1]).get();
+  return (tuple_name == name && tuple_val == value);
+}
+
+CEQL::FormulaToLogicalCEA
+query_to_logical_cea(Catalog& catalog, CEQL::Query& query) {
+  CEQL::FormulaToLogicalCEA visitor = CEQL::FormulaToLogicalCEA(catalog);
+
+  query.where.formula->accept_visitor(visitor);
+  if (!query.select.is_star) {
+    query.select.formula->accept_visitor(visitor);
+  }
+
+  return visitor;
+}
+}
